Move QueryType to persistence module.
