+++ config/src/main/java/com/alibaba/nacos/config/server/service/repository/embedded/DatabaseOperate.java
@@ -18,7 +18,7 @@ package com.alibaba.nacos.config.server.service.repository.embedded;
 
 import com.alibaba.nacos.common.model.RestResult;
 import com.alibaba.nacos.persistence.repository.embedded.EmbeddedStorageContextHolder;
-import com.alibaba.nacos.persistence.repository.embedded.ModifyRequest;
+import com.alibaba.nacos.persistence.repository.embedded.sql.ModifyRequest;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.io.File;
